project folder restructuring
