+++ src/main/java/org/skypro/bank/star/recommendations_service/configuration/RetryConfiguration.java
@@ -1,4 +1,4 @@
-package org.skypro.bank.star.recommendations_service.configuration.statistics;
+package org.skypro.bank.star.recommendations_service.configuration;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.retry.annotation.EnableRetry;
