+++ src/main/java/org/skypro/bank/star/recommendations_service/configuration/TelegramBotConfig.java
@@ -1,6 +1,6 @@
 package org.skypro.bank.star.recommendations_service.configuration;
 
-import org.skypro.bank.star.recommendations_service.service.TelegramRecommendationHandler;
+import org.skypro.bank.star.recommendations_service.telegram.TelegramRecommendationHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
