+++ src/main/java/org/skypro/bank/star/recommendations_service/telegram/command/StartCommand.java
@@ -1,6 +1,6 @@
-package org.skypro.bank.star.recommendations_service.command.telegram;
+package org.skypro.bank.star.recommendations_service.telegram.command;
 
-import org.skypro.bank.star.recommendations_service.service.TelegramMessageFormatter;
+import org.skypro.bank.star.recommendations_service.telegram.formater.TelegramMessageFormatter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
