+++ src/main/java/org/skypro/bank/star/recommendations_service/telegram/command/TelegramCommandDispatcher.java
@@ -1,4 +1,4 @@
-package org.skypro.bank.star.recommendations_service.command.telegram;
+package org.skypro.bank.star.recommendations_service.telegram.command;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
