+++ src/main/java/org/skypro/bank/star/recommendations_service/telegram/command/UnknownCommand.java
@@ -1,6 +1,6 @@
-package org.skypro.bank.star.recommendations_service.command.telegram;
+package org.skypro.bank.star.recommendations_service.telegram.command;
 
-import org.skypro.bank.star.recommendations_service.service.TelegramMessageFormatter;
+import org.skypro.bank.star.recommendations_service.telegram.formater.TelegramMessageFormatter;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.Update;
