+++ src/test/java/org/skypro/bank/star/recommendations_service/telegrambot_tests/TelegramBotIntegrationTest.java
@@ -1,9 +1,9 @@
 package org.skypro.bank.star.recommendations_service.telegrambot_tests;
 
 import org.junit.jupiter.api.Test;
-import org.skypro.bank.star.recommendations_service.command.telegram.TelegramCommandDispatcher;
-import org.skypro.bank.star.recommendations_service.service.TelegramBotFacade;
-import org.skypro.bank.star.recommendations_service.service.TelegramMessageFormatter;
+import org.skypro.bank.star.recommendations_service.telegram.command.TelegramCommandDispatcher;
+import org.skypro.bank.star.recommendations_service.telegram.TelegramBotFacade;
+import org.skypro.bank.star.recommendations_service.telegram.formater.TelegramMessageFormatter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
